% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVirtualenv.R
\name{createVirtualenv}
\alias{createVirtualenv}
\title{Create a Python virtual environment with necessary packages}
\usage{
createVirtualenv()
}
\description{
This function is used to create a Python virtual environment (called "iSSVD") and installs
necessary packages such as pandas, numpy, and scikit-learn to the environment. Once created, the
environment will automatically be used upon loading the package.
}
\details{
If there is an error installing the Python packages, try restarting your computer and running R/RStudio as administrator.
If the virtual environment is not created, the user's default system Python installation will be used.
In this case, the user will need to have the following packages in their main local Python installation:
\itemize{
  \item pandas
  \item matplotlib
  \item seaborn
  \item scikit-learn
  \item numpy
}
Alternatively, the user can use their own virtual environment with reticulate by activating it with
reticulate::use_virtualenv() or a similar function prior to loading RandMVLearn.
}
\examples{
# Create Python virtual environment "iSSVD"
createVirtualenv()
}
