% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics}
\alias{diagnostics}
\title{Evaluating Biclustering Performance}
\usage{
diagnostics(rows, cols, true_rows, true_cols, n = 200L, p = 1000L, D = 2L)
}
\arguments{
\item{rows}{The indices of rows from the original dataset that have been grouped into biclusters by iSSVDR.}

\item{cols}{A nested list of arrays corresponding to the column indices of the variables included in each bicluster for each view (dataset).}

\item{true_rows}{The true indices of rows that define the biclusters in the original dataset. It is a list of arrays, each containing the row indices of a known true bicluster.}

\item{true_cols}{A nested list of arrays containing the true indices of the variables that define the biclusters in each view of the original dataset.}

\item{n}{The total number of samples (rows) in the original dataset. This is used to initialize arrays and to calculate the diagnostics. Need to append a letter L to the integer. Default: 200.}

\item{p}{The total number of variables (columns) in the original dataset. Like n, this is used in the initialization of arrays for the diagnostics. Need to append a letter L to the integer. Default: 1000.}

\item{D}{The number of views or datasets used in the integrative biclustering. Need to append a letter L to the integer. Default: 2.}
}
\value{
A list with the following elements:
\item{Recovery}{ Measures how well the detected clusters match the true clusters. A higher value indicates better recovery of the true cluster structure.}
\item{Relevance}{ Similar to recovery, this metric assesses the relevance of the detected clusters to the true clusters. A higher value indicates that the detected clusters are more relevant to the true cluster structure.}
\item{F-Score}{ The F-Score combines the information of both recovery and relevance into a single score, with higher values indicating better clustering performance.}
\item{False Positives (FPs)}{ The average minimum proportion of false positive elements across all detected clusters. A lower value indicates fewer elements are wrongly included, which is desirable.}
\item{False Negatives (FNs)}{ The average minimum proportion of false negative elements across all detected clusters. A lower value indicates fewer elements are wrongly excluded, which is also desirable.}
}
\description{
The function evaluates the performance of biclustering results by comparing them against true biclusters using various metrics.
}
\details{
The function is designed to evaluate the performance of biclustering results, particularly focusing on the comparison between discovered biclusters and true biclusters. It calculates several diagnostic metrics, including recovery, relevance, F1 score, false positive rate, and false negative rate. Please refer to the main paper \href{https://doi.org/10.1177/09622802221122427}{Zhang, Weijie, et al. "Robust integrative biclustering for multi-view data." Statistical methods in medical research 31.11 (2022): 2201-2216.}  for more details.
}
\examples{
library(iSSVD)
# Generate two views with four biclusters, number of samples 50, number of variables 1000.
mydata = generateData(myseed = 20L) # Uses default settings
X1 = mydata[[1]][[1]][[1]]
X2 = mydata[[1]][[1]][[2]]
Xdata1 = list(X1, X2)

iSSVD_results = iSSVDR(X = Xdata1, myseed = 25L, standr = FALSE, vthr = 0.9, nbicluster = 4L)
sample_clusters = iSSVD_results[[3]][[3]]
variable_clusters = iSSVD_results[[3]][[4]]

true_sample_clusters = mydata[[2]][[1]]
true_variable_clusters = mydata[[3]][[1]]

diag = diagnostics(rows = sample_clusters, cols = variable_clusters, true_rows = true_sample_clusters, true_cols = true_variable_clusters,
n = 200L, p = 1000L, D = 2L)
}
\references{
\href{https://doi.org/10.1177/09622802221122427}{Zhang, Weijie, et al. "Robust integrative biclustering for multi-view data." Statistical methods in medical research 31.11 (2022): 2201-2216.}

\href{https://doi.org/10.1111/j.1467-9868.2010.00740.x}{Meinshausen, Nicolai, and Peter Bühlmann. "Stability selection." Journal of the Royal Statistical Society Series B: Statistical Methodology 72.4 (2010): 417-473.}

\href{https://doi.org/10.1093/bioinformatics/btr322}{Sill, Martin, et al. "Robust biclustering by sparse singular value decomposition incorporating stability selection." Bioinformatics 27.15 (2011): 2089-2097.}
}
