% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{generateData}
\alias{generateData}
\title{Generate simulated data for a given number of clusters}
\usage{
generateData(
  myseed = 25L,
  n = 200L,
  p = 1000L,
  D = 2L,
  rowsize = 50L,
  colsize = 100L,
  numbers = 1L,
  sigma = 0.1,
  nbicluster = 4L,
  orthonm = FALSE
)
}
\arguments{
\item{myseed}{An integer to set a seed. Need to append a letter L to the integer, for example 25L. It's set to a default value of 25.}

\item{n}{An even integer for the number of rows in the generated data. It's set to a default value of 200.}

\item{p}{An integer for the number of the number of columns in the generated data. It's set to a default value of 1000.}

\item{D}{The number of views or data sources. In multi-view matrix factorization, you often have data from multiple sources or views. D controls how many such views are generated. It's set to a default value of 2.}

\item{rowsize}{An integer for the number of rows in each submatrix. It's set to a default value of 50.}

\item{colsize}{An integer for the number of columns in each submatrix. It's set to a default value of 100.}

\item{numbers}{An integer for the number of sets of data matrices to be generated. It's set to a default value of 1.}

\item{sigma}{A number for the standard deviation of the random noise added to the data. It's set to a default value of 0.1.}

\item{nbicluster}{The number of clusters in the row and column spaces for each synthetic dataset. It's set to a default value of 4.}

\item{orthonm}{A boolean flag. If TRUE, The left singular vector and right singular vector are orthogonalized. The default is FALSE.}
}
\value{
The function will return a list of synthetic data matrices. The list has a structure that reflects the different samples and factors in the generated data.
}
\description{
The function generate data matrices with the truth of underlying biclusters that can be used for dianostics.
}
\details{
Please refer to the main paper \href{https://doi.org/10.1177/09622802221122427}{Zhang, Weijie, et al. "Robust integrative biclustering for multi-view data." Statistical methods in medical research 31.11 (2022): 2201-2216.}  for more details.
}
\examples{
# Generate the data list
mydata = generateData(myseed = 25L, n = 200L, p = 1000L, D = 2L, rowsize = 50L, colsize = 100L,
                      numbers = 1L, sigma = 0.1, nbicluster = 4L, orthonm = FALSE)

# Select two data X1, X2 from the generated data list
X1 = mydata[[1]][[1]][[1]]
X2 = mydata[[1]][[1]][[2]]

# Combine two data into a list
Xdata1 = list(X1, X2)
}
\references{
\href{https://doi.org/10.1177/09622802221122427}{Zhang, Weijie, et al. "Robust integrative biclustering for multi-view data." Statistical methods in medical research 31.11 (2022): 2201-2216.}

\href{https://doi.org/10.1111/j.1467-9868.2010.00740.x}{Meinshausen, Nicolai, and Peter Bühlmann. "Stability selection." Journal of the Royal Statistical Society Series B: Statistical Methodology 72.4 (2010): 417-473.}

\href{https://doi.org/10.1093/bioinformatics/btr322}{Sill, Martin, et al. "Robust biclustering by sparse singular value decomposition incorporating stability selection." Bioinformatics 27.15 (2011): 2089-2097.}
}
