% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatMapR.R
\name{plotHeatMapR}
\alias{plotHeatMapR}
\title{Plot heatmap using generated and results from iSSVDR}
\usage{
plotHeatMapR(X = Xdata, Rows = myRows, Cols = myCols, D = 2L, nbicluster = 4L)
}
\arguments{
\item{X}{A list of matrices to be visualized. It's assumed that this list contains multiple views or data sources. The length of this list should be equal to the value of D, representing the number of views. Each matrix in the list corresponds to one view of the data.}

\item{Rows}{A list of indices that represent the sample clusters from iSSVDR.}

\item{Cols}{A list of indices that represent the variable clusters from iSSVDR.}

\item{D}{This parameter represents the number of views or data sources in the synthetic data. It's set to 2 by default, but you can change it to match the actual number of views in your data.}

\item{nbicluster}{The number of biclusters to visualize. This parameter specifies how many biclusters should be included in the heatmap visualization. It defaults to 4, but you can adjust it to visualize a different number of biclusters.}
}
\value{
The result of this function is the display of two heatmaps side by side, each representing a different dimension of the data.
}
\description{
The plotHeatMapR function takes a data matrix, organizes it into biclusters, and creates and displays two side-by-side heatmaps to visualize the data's structure and relationships within two dimensions.
}
\details{
Please refer to the main paper \href{https://doi.org/10.1177/09622802221122427}{Zhang, Weijie, et al. "Robust integrative biclustering for multi-view data." Statistical methods in medical research 31.11 (2022): 2201-2216.} for more details.
}
\examples{
# Generate data and perform integrative biclustering using iSSVDR()
mydata = generateData(myseed = 20L) # Uses default settings
X1 = mydata[[1]][[1]][[1]]
X2 = mydata[[1]][[1]][[2]]
Xdata1 = list(X1, X2)

iSSVD_results = iSSVDR(X = Xdata1, myseed = 25L, standr = FALSE, vthr = 0.9, nbicluster = 4L)

# Plot HeatMap using the results from iSSVDR
myRows = iSSVD_results[[3]][[3]]
myCols = iSSVD_results[[3]][[4]]
myRows <- lapply(myRows, as.integer)
myCols <- lapply(myCols, lapply, as.integer)

plotHeatMapR(X = Xdata1, Rows = myRows, Cols = myCols, D = 2L, nbicluster = 4L)
}
\references{
\href{https://doi.org/10.1177/09622802221122427}{Zhang, Weijie, et al. "Robust integrative biclustering for multi-view data." Statistical methods in medical research 31.11 (2022): 2201-2216.}

\href{https://doi.org/10.1111/j.1467-9868.2010.00740.x}{Meinshausen, Nicolai, and Peter Bühlmann. "Stability selection." Journal of the Royal Statistical Society Series B: Statistical Methodology 72.4 (2010): 417-473.}

\href{https://doi.org/10.1093/bioinformatics/btr322}{Sill, Martin, et al. "Robust biclustering by sparse singular value decomposition incorporating stability selection." Bioinformatics 27.15 (2011): 2089-2097.}
}
