% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summary_table}
\alias{summary_table}
\title{Make summary table for clinical data}
\usage{
summary_table(data, res, ...)
}
\arguments{
\item{data}{Input clinical data. Can use data("clinical_data") to use example data that contains 120 rows and 4 variables: "Age", "Sex", "crp", "ddimer".}

\item{res}{Input results directly generated from the iSSVDR() function.}

\item{...}{Any variables you want to be shown in the summary table.}
}
\value{
A summary table that displays and compares summary statistics across sample clusters, and give p-values.
}
\description{
The function allows the user to make a summary table based on clinical data and sample clusters get from iSSVDR() function.
}
\details{
summary_table returns a summary table showing you the summary statistics (N, Mean, SD) and p-value from Kruskal-Wallis rank sum test (for continuous variables)/ Pearson’s Chi-squared test (for categorical variables). This function depends on the gtsummary package.
}
\examples{
library(iSSVD)
# Call example data
data("clinical_data") # Clinical data has no association with clusters, this is just for demonstrations.

# Generate three views with four biclusters, number of samples 120, number of variables 1000.
mydata = generateData(n = 120L, p = 1000L, D = 3L, rowsize = 30L, colsize = 100L,
                      numbers = 1L, sigma = 0.1, nbicluster = 4L, orthonm = FALSE)
Xdata1 <- list()
for (i in 1:3) {
  Xdata1[[i]] <- mydata[[1]][[1]][[i]]
}

res4 = iSSVDR(X = Xdata1, standr = FALSE, pointwise = TRUE, steps = 100, size = 0.5,
              vthr = 0.9, ssthr = c(0.6, 0.65), nbicluster = 4, rows_nc = FALSE, cols_nc = FALSE,
              col_overlap = FALSE, row_overlap = FALSE, pceru = 0.8, pcerv = 0.8, merr = 0.0001,
              iters = 100, assign_unclustered_samples = FALSE)

# Generate Table
summary_table(clinical_data, res4, Age, Sex, crp, ddimer)
}
\references{
\href{https://doi.org/10.1177/09622802221122427}{Zhang, Weijie, et al. "Robust integrative biclustering for multi-view data." Statistical methods in medical research 31.11 (2022): 2201-2216.}

\href{https://doi.org/10.1111/j.1467-9868.2010.00740.x}{Meinshausen, Nicolai, and Peter Bühlmann. "Stability selection." Journal of the Royal Statistical Society Series B: Statistical Methodology 72.4 (2010): 417-473.}

\href{https://doi.org/10.1093/bioinformatics/btr322}{Sill, Martin, et al. "Robust biclustering by sparse singular value decomposition incorporating stability selection." Bioinformatics 27.15 (2011): 2089-2097.}
}
